Remove Plots backend components
