+++ Project.toml
@@ -14,7 +14,6 @@ Memento = "f28f55f0-a522-5efc-85c2-fe41dfb9b2d9"
 NLopt = "76087f3c-5699-56af-9a33-bf431cd00edd"
 NetworkLayout = "46757867-2c16-5918-afeb-47bfcb05e46a"
 OMEinsum = "ebe7aa44-baf0-506c-a96f-8464559b3922"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PowerModels = "c36e90e8-916a-50a6-bd94-075b64ef4655"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 RecursiveArrayTools = "731186ca-8d62-57ce-b412-fbd966d074cd"
