+++ test/runtests.jl
@@ -2,7 +2,7 @@ using PowerPlots
 using Test
 
 using PowerModels
-import Ipopt
+# import Ipopt
 
 using Memento
 using Memento.TestUtils
@@ -12,23 +12,8 @@ PowerModels.logger_config!("error") # silence PowerModels logger while testing
 data = PowerModels.parse_file("$(joinpath(dirname(pathof(PowerModels)), ".."))/test/data/matpower/case5.m")
 
 @testset "PowerPlots.jl" begin
-    # Write your own tests here.
 
-    @testset "Plots Backend" begin
-        graph = build_graph_network(data; exclude_sources=true)
-        plot_network(data)
-        plot_network!(data)
 
-        plot_network_status!(graph)
-        plot_system_voltage(data)
-
-        result = run_opf(data, DCPPowerModel, Ipopt.Optimizer)
-        PowerModels.update_data!(data, result["solution"])
-        plot_power_flow(data)
-
-        # these functions work if we got this far
-        @test true # these functions didn't error
-    end
 
     @testset "VegaLite Backend" begin
         ## Test vegalite plotting
